Remove old hacky layers, reinstate icons
