+++ RGBMasterWinUI/RGBMasterWinUI (Package)/Images/SplashScreen.scale-100.png
