+++ RGBMasterWinUI/RGBMasterWinUI (Package)/Package.appxmanifest
@@ -12,7 +12,7 @@
     Version="2.6.0.0" />
 
   <Properties>
-    <DisplayName>RGBMasterWinUI (Package)</DisplayName>
+    <DisplayName>RGBMaster</DisplayName>
     <PublisherDisplayName>Lochness Monster</PublisherDisplayName>
     <Logo>Images\StoreLogo.png</Logo>
   </Properties>
@@ -31,12 +31,12 @@
       Executable="$targetnametoken$.exe"
       EntryPoint="$targetentrypoint$">
       <uap:VisualElements
-        DisplayName="RGBMasterWinUI (Package)"
-        Description="RGBMasterWinUI (Package)"
+        DisplayName="RGBMaster"
+        Description="RGBMaster"
         BackgroundColor="transparent"
         Square150x150Logo="Images\Square150x150Logo.png"
         Square44x44Logo="Images\Square44x44Logo.png">
-        <uap:DefaultTile Wide310x150Logo="Images\Wide310x150Logo.png" />
+        <uap:DefaultTile Wide310x150Logo="Images\Wide310x150Logo.png"  Square71x71Logo="Images\SmallTile.png" Square310x310Logo="Images\LargeTile.png"/>
         <uap:SplashScreen Image="Images\SplashScreen.png" />
       </uap:VisualElements>
     </Application>
