+++ RGBMasterWinUI/RGBMasterWinUI/MainWindow.xaml
@@ -14,5 +14,74 @@
         <Button x:Name="myButton" Click="myButton_Click">Click Me</Button>
     </StackPanel>-->
 
-    <local:RGBMasterUserControl></local:RGBMasterUserControl>
+    <Grid Name="ContainerGrid" Loaded="ContainerGrid_Loaded">
+        <Grid.Resources>
+            <ResourceDictionary>
+                <ResourceDictionary.ThemeDictionaries>
+                    <ResourceDictionary x:Key="Dark">
+                        <!--<SolidColorBrush x:Key="ButtonBackground" Color="#00a69a"/>
+                    <SolidColorBrush x:Key="ButtonBackgroundPointerOver" Color="#4cc0b8"/>
+                    <SolidColorBrush x:Key="ButtonBackgroundPressed" Color="#00635c"/>
+                    <SolidColorBrush x:Key="PivotHeaderItemSelectedPipeFill" Color="#00a69a" />
+                    
+                    <SolidColorBrush x:Key="TextControlBorderBrush" Color="#00a69a"/>
+                    <SolidColorBrush x:Key="TextControlBorderBrushPointerOver" Color="#4cc0b8"/>-->
+                    </ResourceDictionary>
+                </ResourceDictionary.ThemeDictionaries>
+            </ResourceDictionary>
+        </Grid.Resources>
+
+        <VisualStateManager.VisualStateGroups>
+            <VisualStateGroup>
+                <VisualState x:Name="None">
+                    <VisualState.StateTriggers>
+                        <AdaptiveTrigger MinWindowWidth="0" />
+                    </VisualState.StateTriggers>
+                    <VisualState.Setters>
+                        <Setter Target="MainNavigationView.IsPaneOpen" Value="False" />
+                    </VisualState.Setters>
+                </VisualState>
+
+                <VisualState x:Name="Compact">
+                    <VisualState.StateTriggers>
+                        <AdaptiveTrigger MinWindowWidth="940" />
+                    </VisualState.StateTriggers>
+                    <VisualState.Setters>
+                        <Setter Target="MainNavigationView.IsPaneOpen" Value="True" />
+                    </VisualState.Setters>
+                </VisualState>
+
+            </VisualStateGroup>
+        </VisualStateManager.VisualStateGroups>
+
+
+
+        <NavigationView x:Name="MainNavigationView" PaneDisplayMode="Left" IsBackButtonVisible="Collapsed" SelectionChanged="MainNavigationView_SelectionChanged" Loaded="NavigationView_Loaded">
+            <NavigationView.PaneHeader>
+                <StackPanel Orientation="Horizontal">
+                    <Image Source="Assets/128.png" Width="23" Height="23" Margin="4 0 4 0" />
+                    <TextBlock VerticalAlignment="Center" Text="RGB Master" FontWeight="SemiBold" FontSize="16" />
+                    <TextBlock Text="{x:Bind CurrentEffectName, Mode=OneWay}" />
+                </StackPanel>
+            </NavigationView.PaneHeader>
+            <NavigationView.MenuItems>
+                <NavigationViewItem Content="Devices" x:Name="DevicesNavigationItem" Tag="DevicesPage">
+                    <NavigationViewItem.Icon>
+                        <FontIcon Glyph="&#xEA80;" FontFamily="Segoe MDL2 Assets"/>
+                    </NavigationViewItem.Icon>
+                </NavigationViewItem>
+                <NavigationViewItem Content="Effects" x:Name="Effects" Tag="EffectsPage">
+                    <NavigationViewItem.Icon>
+                        <FontIcon Glyph="&#xE790;" FontFamily="Segoe MDL2 Assets" />
+                    </NavigationViewItem.Icon>
+                </NavigationViewItem>
+            </NavigationView.MenuItems>
+            <ScrollViewer>
+                <StackPanel>
+                    <Frame x:Name="MainAppContentFrame" Padding="12,0,12,0" IsTabStop="True"
+                   NavigationFailed="MainAppContentFrame_NavigationFailed"/>
+                </StackPanel>
+            </ScrollViewer>
+        </NavigationView>
+    </Grid>
 </Window>
